finish renaming shiny -> ui
