+++ .flake8
@@ -1,5 +1,5 @@
 [flake8]
-exclude = .git,.venv,.venv-*,__pycache__,.local-sessions
+exclude = .git,.venv,.venv-*,__pycache__,.local-sessions,dist
 
 # Config recommended by black:
 # https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#bugbear
