+++ dp_wizard/app.py
@@ -1,4 +1,4 @@
-from dp_wizard.shiny import make_app
+from dp_wizard.ui import make_app
 from dp_wizard.utils.argparse_helpers import get_cli_info
 
 app = make_app(get_cli_info())
