+++ pyproject.toml
@@ -47,7 +47,7 @@ ignore = ["**/no-tests/"]
 reportPrivateImportUsage = false
 # TODO: Expand strict type checking
 strict = [
-    "dp_wizard/shiny/__init__.py",
+    "dp_wizard/ui/__init__.py",
     "dp_wizard/utils/argparse_helpers.py",
     "dp_wizard/utils/config.py",
     "dp_wizard/utils/csv_helper.py",
