+++ tests/apps/app_sample.py
@@ -1,4 +1,4 @@
-from dp_wizard.shiny import make_app
+from dp_wizard.ui import make_app
 from dp_wizard.utils.argparse_helpers import cli_info_defaults
 
 app = make_app(
